requirements.txt
flask
torch
facenet-pytorch
numpy
Pillow
scikit-learn
joblib
gunicorn
python-dotenv
matplotlib
